created scraper for selftape audition
