+++ src/pages/background/googleapi.ts
@@ -1,6 +1,7 @@
 import { GoogleSpreadsheet } from "google-spreadsheet";
 import { getAccessToken } from "./authorize";
-import { SHEET_NAME, Audition, AABrowserReq, IFoundFile, TEMPLATE_ID } from ".";
+import { SHEET_NAME, TEMPLATE_ID } from ".";
+import { AABrowserReq, IFoundFile } from "@src/types";
 
 const CLIENT_ID =
   "635620722112-iokrike3aui2lacke3ncoulooforlm81.apps.googleusercontent.com";
