+++ src/pages/content/sendToGoogleSheets.ts
@@ -1,38 +1,17 @@
 import $ from "jquery";
+import {
+  getPersonFromHTML,
+  getProjectType,
+  parseEntryFromHtml,
+} from "./helperFunctions";
+import { LAST_UPDATED_FN } from "@src/constants";
 
 const AA_MAIN_URL = "https://actorsaccess.com";
 const projectURL = AA_MAIN_URL + $(".cart_role_breakdown").attr("href");
 const breakdownCell = $(".roleItem").html();
 const submitButton = $("#cartsubmit");
 const submitForm = $("form.submit-container");
-export function parseEntryFromHtml(
-  html: string,
-  tag: string,
-  word: string,
-): string {
-  const wordStart = html.indexOf(word);
-  let subString = html.slice(wordStart);
-  const brTagIndex = subString.indexOf("<br>", word.length);
-  subString = subString.slice(0, brTagIndex);
-  return subString.slice(word.length + tag.length + 1).trim();
-}
 
-export function getPersonFromHTML(html: string, word: string): string {
-  const wordStart = html.indexOf(word);
-  let subString = html.slice(wordStart);
-  const brTagIndex = subString.indexOf("<br>", word.length);
-  subString = subString.slice(0, brTagIndex);
-  subString = subString.slice(word.length + 1);
-  return subString;
-}
-
-export function getProjectType(html: string) {
-  const br = "<br>";
-  const firstBreak = html.indexOf(br);
-  let substring = html.slice(firstBreak + br.length);
-  const secondBreak = substring.indexOf("<");
-  return substring.slice(0, secondBreak);
-}
 submitButton.html("Submit <span class=underline-text> and Track </span");
 submitForm.one("submit", async (e) => {
   e.preventDefault();
@@ -55,9 +34,6 @@ submitForm.one("submit", async (e) => {
       projectType = getProjectType(leftTable.html());
       casting = getPersonFromHTML(rightTable.html(), "Casting Director:");
     }).then(() => {
-      const lastUpdatedFunction =
-        '=IF(COUNTA($A$1:$G$1)=0,"",iferror($A$1:$G$1+"x",today()))';
-
       const audition = {
         orderNo: "1",
         submittedDate: new Date().toLocaleDateString(),
@@ -66,7 +42,7 @@ submitForm.one("submit", async (e) => {
         castingDirector: casting,
         projectType,
         status: "Submitted",
-        lastUpdated: lastUpdatedFunction,
+        lastUpdated: LAST_UPDATED_FN,
       };
 
       const sendMessage = browser.runtime.sendMessage({ audition: audition });
